LayoutTests/workers/sab/postMessage-transfer-type-error-expected.txt

Checks that worker.postMessage rejects SharedArrayBuffers in the transfer list

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS postMessageTransferSharedArrayBuffer() threw exception TypeError: Cannot transfer a SharedArrayBuffer.
PASS successfullyParsed is true

TEST COMPLETE
